Poprawki w algorytmie, dodano parametr wielkosci wszystkich plikow uzytkownikow.

+++ src/main/java/loadBalancer/FileMockLoadBalancer.java
@@ -7,7 +7,7 @@ import lombok.Data;
 public class FileMockLoadBalancer implements Comparable<FileMockLoadBalancer> {
 
     private FileMock fileMock;
-    private int priority;
+    private double priority;
 
     FileMockLoadBalancer(FileMock fileMock) {
         this.fileMock = fileMock;
